Updated zero handling logLikStat in functions.R
Before if add_one was one we always added one to all zero simulation values. The problem with this is that if the data was also zero, this is artificially moving the model away from the data and lowering the likelihood. In addition, having data values of zero is also a problem for some likelihood functions. Now, we do something slightly more nuanced.
*  If the model and data are both zero at a timepoint, we assign a likelihood of 1.
* If only the model is zero but data is non-zero, if add_one is TRUE we add one to model value and calculate likelihood.
* If the model is non-zero, then the likelihood is evaluated as normal UNLESS the function is lognormal, and then we add one to the observed data