Unit test `Parameters` exceptions in init
Initialized the unit tests for the `gempyor.parameters.Parameters` class
with fixtures for the exceptions raised by the class' constructor. Left
a note about the inability to reach one of the `ValueError`s, will have
to revisit, either with a way to reach that exception or refactor the
dead code.