Added unit tests for `Statistic.apply_scale`
Added unit tests for `apply_scale` method including a new factory that
produces an input set with a 'scale' config. Fixed a bug where the scale
function was not applied even if provided. This is a *breaking* change,
but doesn't affect currently existing test suite, need to see if this
affects any currently existing config files.