Bug fix `compute_all_multioutcomes`
Added a bug fix to `compute_all_multioutcomes` to look for if a sourced
outcome exists and has not been processed yet and if so skip that
outcome for now and re-add it to the end of the outcomes to process.
Required for the `test_outcomes_pcomp` and
`test_outcomes_pcomp_read_write` tests to pass. Also minor edits to
`config_mc_selection.yml` to add the required 'compartments'/'seir'
sections.