Expand scalar likelihoods in `Statistics.llik`
Manually expand likelihoods to match `gt_data.shape` if they are scalars
(rmse or absolute_error). `xarray` used to do this because it would
infer `np.float64` was not a scalar based on the absense of the
'__array_namespace__' attr. This changed somewhere between v2.0.2 and
v2.1.3, maybe in numpy/numpy#26889 but not certain (PR description would
suggest otherwise).