Internal helper to calculate resources from size
Added `_job_resources_from_size_and_inference` to calculate job
resources from job size and inference type along with corresponding docs
and tests. This is a temporary solution since a method like this should
be added to GH-402/GH-432 rather than hard coded in `gempyor.batch`.
Simpler to do this way for MVP's sake rather than trying to abstract
code in a way that won't be reusable in the future.