Add `BatchSystem.submit_command` method
Added a the `submit_command` method to `gempyor.batch.BatchSystem` to
act as a useful wrapper around the `submit` abstract method. Instead of
taking a full script only takes the command to be ran. Implemented said
method for `SlurmBatchSystem` to render an sbatch script from the yet to
be created `sbatch_submit_command.bash.j2` template. Yet to add unit
tests.