Update R inference acceptance decision
I merged in and improved upon the changes suggested by Shaun in old PR 379 to use a function for single metropolis hasting acceptance steps used in the MCMC algorithm, instead of different functions for chimeric vs global decision (which could very occasionally give different results due to having u<a instead of u<=a in the chimeric decision). I also made the interateAccept function return both the acceptance decision and acceptance probability, or optionally, only the ratio. Now we don't need to separately calculate this ratio in the script (flepimop-inference-slot) in order to record it along with the likelihoods for diagnostic purposes. I made some updates to the inference tests (test-iterateAccept) for this function to match these edits. In addition, I could not figure out the logic for one of the tests in there so I just made a more transparent version of it.
Separately, I made some minor updates to the fake data for running inference checks with the sample 2 population tutorial exampls.