Add `gempyor.utils.ConfigurationWarning`
Added a custom warning for potential configuration file issues and
utilized this new warning in
`gempyor.initial_conditions._file_or_folder_draw.FileOrFolderDrawInitialConditions`
class when validating a configuration file if that file sets a
non-default value for `proportional_ic` when the method chosen does not
use it. This situtation likely indicates that the non-default value
needs to be removed or maybe that the method chosen was picked
incorrectly by the user.