Fix multiprocessing issues with custom plugins
When using initial conditions custom plugins, like ones that can be
inferred, the module has to be related when the multiprocessing start
method is 'spawn' or 'forkserver' because the child process does not get
a copy of the parent process' memory and therefore does not have the
dynamically loaded plugin available at start. This can lead to an unsual
'ModuleNotFound' exception with a traceback relating to deserializing
the pickled memory state. Luckily, the fix is simple, just need to
reload the custom plugin when the child process starts. This should be
kept in mind, and likely refactored when addressing #421.