Modify 'usa state level' example to infer ICs
Modified the example contained in `examples/simple_usa_statelevel/` to
infer the initial conditions with a custom plugin. This custom initial
conditions plugin is very simple, just figuring out what proportion of
the population to place into the susceptible vs the infected
compartments. However, this is a great way to introduce users into
inferable initial conditions.