Modify R inference to handle IC plugins
Made minor adjustments to the `inference` R package to support initial
condition plugins. Mostly adjusting logic around which code path to go
down based on the `method` field of the `initial_conditions` section.

Also fixed a bug in the `flepimop-inference-main` command where it was
trying to call the `flepimop-inference-slot` command with `R` rather
than directly invoking it.