Add `DistributionABC.support` property
Added a `support` property to the `DistributionABC` class which returns
a tuple of two numbers representing the upper and lower bounds
respectively. This property pulls said bounds from either the
`_lower/upper_bound` attributes if defined by the implementation or
`_lower/upper` properties if overridden by the implementation. Added to
the current examples to demonstrate usage. Minor modifications to the
commond distributions test was required.